requirements.txt
gradio==5.23.2
torch
torchvision
Pillow
requests